+++ sea-orm-cli/src/commands/generate.rs
@@ -31,6 +31,8 @@ pub async fn run_generate_command(
             lib,
             model_extra_derives,
             model_extra_attributes,
+            enum_extra_derives,
+            enum_extra_attributes,
             seaography,
         } => {
             if verbose {
@@ -180,6 +182,8 @@ pub async fn run_generate_command(
                 serde_skip_hidden_column,
                 model_extra_derives,
                 model_extra_attributes,
+                enum_extra_derives,
+                enum_extra_attributes,
                 seaography,
             );
             let output = EntityTransformer::transform(table_stmts)?.generate(&writer_context);